fix: let `Table` call first `getPageData` instead of using `initialData` in `ListPartner` and `ListUser` views
This is due to a bug that causes the `Table` component to call
`getPageData` infinitely when both `initialData` and `getPageData` are
passed to it.

The bug happens because the `Table` component calls `getPageData` from the
`watch` hook, which is triggered when `initialData` changes. This causes
`initialData` to be updated, which triggers the `watch` hook again, and so
on.